Changed input variables to command line args
I changed the general variables to use command line args instead. Its now a little easier to use from different computers. The names of the VSWR files are stored as key value pairs in another file that gets read. I've put this file in the Inuvik trips folder of the owncloud.