Presentation and conversion to pandas dataframes
Modified functions in dataset_operations module to work with pandas
dataframes instead of numpy arrays with dtypes. Removed print statements
in retrieve_data module. Created a load_data script for presentation,
which can be passed the corresponding site, date, and data type
information to grab the data corresponding to that information. Updated
presentation and Jupyter notebook for testing.