load_data.py changes for more generality
load_data.py was changed to:
- take in an index for the metadata from the site_file_metadata
dataframe, instead of the site, date, and data_type (which may not be
unique because there can be pre-fix 'interim data' from the same day as
the post-fix 'non-interim data'.
- create a colour_dictionary for plotting for the channels available
from the mapping file.
- take in some keys from the mapping file that are not actually channel
names but just more information about that test data, ie.
'atten, 'estimate_intf' and 'atten_file'.
- incorporate attenuation data for pm-paths.
- only combine the channels into arrays if there is individual channel data
available for this
- improved print statements for using in Jupyter notebooks.